==> Notebooks/1_Clean_AEMET.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# Especifica la ruta al archivo CSV
path = r'C:\\Users\\joset\\OneDrive\\Documentos\\Master Big Data UCM\\18-TFN\NaturgyProject\\data'
# Lee el archivo CSV en un DataFrame de pandas
AEMET_Previsiones_original = pd.read_csv(path + '\\Originales\\AEMET_Previsiones.csv', encoding = 'latin1')
AEMET_Realidad_original = pd.read_csv(path + '\\Originales\\AEMET_Realidad.csv', encoding = 'latin1')

AEMET_Realidad = AEMET_Realidad_original.copy()
AEMET_Previsiones = AEMET_Previsiones_original.copy()

# Crear una nueva columna FECHA agrupando DIA/MES/ANYO como en el anterior dataframe

AEMET_Realidad['DIA'] = AEMET_Realidad['DIA'].astype(str).str.zfill(2)
AEMET_Realidad['MES'] = AEMET_Realidad['MES'].astype(str).str.zfill(2)

AEMET_Realidad['FECHA'] = AEMET_Realidad['DIA'] + '/' + AEMET_Realidad['MES'] + '/' + AEMET_Realidad['ANYO'].astype(str)
AEMET_Realidad['FECHA'] = pd.to_datetime(AEMET_Realidad['FECHA'], format='%d/%m/%Y')

AEMET_Realidad.head()

print(AEMET_Realidad.dtypes)

AEMET_Realidad[['MES', 'DIA']] = AEMET_Realidad[['MES', 'DIA']].astype(int)

# Las filas de 3168C son duplicados de 3168D. 3168C solo llega hasta 2014
AEMET_Realidad = AEMET_Realidad[AEMET_Realidad['INDICATIVO'] != '3168C']

# 1249I No tiene valores despúes de 2014 posiblemente sea el anterior a 1249x porque hay meses de 2014 que parece que se repiten las mediciones.
# Como vamos a utilizar a partir de 2016 no hay problema
AEMET_Realidad = AEMET_Realidad[AEMET_Realidad['INDICATIVO'] != '1249I']

# 1111 Parece que tiene valores duplicados con 1111X y solo llega hasta 2014
AEMET_Realidad = AEMET_Realidad[AEMET_Realidad['INDICATIVO'] != '1111']

# 8416 Parece que tiene valores duplicados con 8416Y y solo llega hasta 2014
AEMET_Realidad = AEMET_Realidad[AEMET_Realidad['INDICATIVO'] != '8416']

# 9434 Parece que tiene valores duplicados con 9434P y solo llega hasta 2014
AEMET_Realidad = AEMET_Realidad[AEMET_Realidad['INDICATIVO'] != '9434']



# Las filas con el INDICATIVO 1387 y LATITUD 433700 correspondientes a mes de marzo del 16 están 'duplicadas'
AEMET_Realidad = AEMET_Realidad[AEMET_Realidad['LATITUD'] != 433700]

# Eliminar las filas donde INDICATIVO es 1387E, LONGITUD es 822192 y LATITUD es nulo. Son duplicidades en fechas detectadas en 2015
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1387E') &
                                                    (AEMET_Realidad['LONGITUD'] == 822192) &
                                                    (AEMET_Realidad['LATITUD'].isnull())].index)

# El observatorio 8178D tiene valores duplicados en octubre del 22 cuando cambia la lat y long
# primero hay que borrar los duplicados (que tienen max y min varcia) y luego aunar las coordenadas
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '8178D') &
                                                    (AEMET_Realidad['LONGITUD'] == 151462) &
                                                    (AEMET_Realidad['LATITUD'] == 390021) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MES'] == 10) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '8178D', ['LATITUD', 'LONGITUD']] = [390021, 151462]

# El observatorio 8175 tiene valores duplicados en diciembre del 16 cuando cambia la lat y long
# primero hay que borrar los duplicados y luego aunar las coordenadas
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '8175') &
                                                    (AEMET_Realidad['LONGITUD'] == 151222) &
                                                    (AEMET_Realidad['LATITUD'] == 385715) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 12)].index)

# El observatorio 8175 tiene valores duplicados en octubre del 16 cuando cambia la lat y long
# primero hay que borrar los duplicados y luego aunar las coordenadas
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '8175') &
                                                    (AEMET_Realidad['LONGITUD'] == 151452) &
                                                    (AEMET_Realidad['LATITUD'] == 385706) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 10)].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '8175') &
                                                    (AEMET_Realidad['LONGITUD'] == 151222) &
                                                    (AEMET_Realidad['LATITUD'] == 385715) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 9)].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '8175', ['LATITUD', 'LONGITUD']] = [385715, 151232]

dias = [8,9,10,11,12]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '2444') &
                                                    (AEMET_Realidad['LONGITUD'] == 440472) &
                                                    (AEMET_Realidad['LATITUD'] == 403933) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['DIA'].isin(dias)) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '2444', ['LATITUD', 'LONGITUD']] = [403933, 440472]

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '0201D', ['LATITUD', 'LONGITUD']] = [412326, 212001]

dias = [8,9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '2331') &
                                                    (AEMET_Realidad['LONGITUD'] == 337172) &
                                                    (AEMET_Realidad['LATITUD'] == 422125) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['DIA'].isin(dias)) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '2331', ['LATITUD', 'LONGITUD']] = [422125, 337172]

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '1082', ['LATITUD', 'LONGITUD']] = [431753, 254232]

dias = [11,12,13,14,15,16]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'BILBAO/AEROPUERTO') &
                                                    (AEMET_Realidad['INDICATIVO'] == '1082') &
                                                    (AEMET_Realidad['LONGITUD'] == 254232) &
                                                    (AEMET_Realidad['LATITUD'] == 431753) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['DIA'].isin(dias)) &
                                                    (AEMET_Realidad['MES'] == 5)].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '1082', 'NOMBRE'] = 'BILBAO AEROPUERTO'

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1428') &
                                                    (AEMET_Realidad['LONGITUD'] == 824382) &
                                                    (AEMET_Realidad['LATITUD'].isnull())].index)

mes = [3,4,5,7,8,9,10]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1495') &
                                                    (AEMET_Realidad['LONGITUD'] == 837262) &
                                                    (AEMET_Realidad['LATITUD'].isnull()) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'].isin(mes))].index)

mes = [1,2,4,5,6,7,9,11,12]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1495') &
                                                    (AEMET_Realidad['LONGITUD'] == 837262) &
                                                    (AEMET_Realidad['LATITUD'].isnull()) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'].isin(mes))].index)

dias = [1,2,3,4,5]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1428') &
                                                    (AEMET_Realidad['NOMBRE'] == 'SANTIAGO DE COMPOSTELA/LAVACOLLA') &
                                                    (AEMET_Realidad['LONGITUD'] == 824382) &
                                                    (AEMET_Realidad['LATITUD'] == 425317) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 5) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [14,15,16,17,18,19,30]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1428') &
                                                    (AEMET_Realidad['NOMBRE'] == 'SANTIAGO DE COMPOSTELA/LAVACOLLA') &
                                                    (AEMET_Realidad['LONGITUD'] == 824382) &
                                                    (AEMET_Realidad['LATITUD'] == 425317) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 4) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

dias2 = [19,20,21,22,23,24]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1428') &
                                                    (AEMET_Realidad['NOMBRE'] == 'SANTIAGO DE COMPOSTELA/LAVACOLLA') &
                                                    (AEMET_Realidad['LONGITUD'] == 824382) &
                                                    (AEMET_Realidad['LATITUD'] == 425317) &
                                                    (AEMET_Realidad['ANYO'] == 2024) &
                                                    (AEMET_Realidad['MES'] == 6) &
                                                    (AEMET_Realidad['DIA'].isin(dias2))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '1428', 'NOMBRE'] = 'SANTIAGO DE COMPOSTELA/LABACOLLA'

dias = [8,9,10,11,12]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1505') &
                                                    (AEMET_Realidad['LONGITUD'] == 727272) &
                                                    (AEMET_Realidad['LATITUD'] == 430640) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1505') &
                                                    (AEMET_Realidad['LONGITUD'] == 727272) &
                                                    (AEMET_Realidad['LATITUD'] == 430640) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 8) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '1505', 'LATITUD'] = 430641

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1549') &
                                                    (AEMET_Realidad['LATITUD'] == 425600) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 5)].index)

dias = [9,10,11,12,13,14]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '2422') &
                                                    (AEMET_Realidad['LONGITUD'] == 445152) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 8) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [8,9,10,11,12]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '2422') &
                                                    (AEMET_Realidad['LONGITUD'] == 445152) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '2422', 'LONGITUD'] = 445162

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '3129') &
                                                    (AEMET_Realidad['LONGITUD'] == 333202) &
                                                    (AEMET_Realidad['LATITUD'].isnull())].index)

dias = [2,3,4,5,6]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '3195') &
                                                    (AEMET_Realidad['NOMBRE'] == 'MADRID, RETIRO')&
                                                    (AEMET_Realidad['ANYO'] == 2018) &
                                                    (AEMET_Realidad['MES'] == 6) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [10,11,12,13,14]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '3195') &
                                                    (AEMET_Realidad['NOMBRE'] == 'MADRID, RETIRO')&
                                                    (AEMET_Realidad['ANYO'] == 2020) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '3195', 'NOMBRE'] = 'MADRID, RETIRO'

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '4121') &
                                                    (AEMET_Realidad['LONGITUD'] == 355132) &
                                                    (AEMET_Realidad['LATITUD'] == 389900)].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '5402') &
                                                    (AEMET_Realidad['DIA'] == 27) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['LONGITUD'] == 450462) &
                                                    (AEMET_Realidad['LATITUD'] == 375039)].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '5402', ['LATITUD', 'LONGITUD']] = [375056, 450482]

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '5973', 'NOMBRE'] = 'CADIZ,OBS.'

dias1 = [9,10,11,12,13,14]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9170') &
                                                    (AEMET_Realidad['NOMBRE'] == 'LOGROÑO, AEROPUERTO')&
                                                    (AEMET_Realidad['ANYO'] == 2021) &
                                                    (AEMET_Realidad['MES'] == 4) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9170', 'NOMBRE'] = 'LOGROÑO, AEROPUERTO'

dias = [29,30]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9898') &
                                                    (AEMET_Realidad['LONGITUD'] == 19322) &
                                                    (AEMET_Realidad['LATITUD'] == 420504) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 9) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [1,2,3]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9898') &
                                                    (AEMET_Realidad['LONGITUD'] == 19322) &
                                                    (AEMET_Realidad['LATITUD'] == 420504) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 10) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

dias2 = [9,10,11,12]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9898') &
                                                    (AEMET_Realidad['LONGITUD'] == 19322) &
                                                    (AEMET_Realidad['LATITUD'] == 420504) &
                                                    (AEMET_Realidad['ANYO'] == 2021) &
                                                    (AEMET_Realidad['MES'] == 4) &
                                                    (AEMET_Realidad['DIA'].isin(dias2))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9898', 'NOMBRE'] = 'HUESCA, AEROPUERTO'

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9898', ['LATITUD', 'LONGITUD']] = [420504, 19322]

dias = [22,23,24,25,26]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '0016A') &
                                                    (AEMET_Realidad['LONGITUD'] == 109491) &
                                                    (AEMET_Realidad['LATITUD'] == 410842) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 9) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '0016A', ['LATITUD', 'LONGITUD']] = [410842, 109491]

dias1 = [11,12,13,14,15]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1024E') &
                                                    (AEMET_Realidad['NOMBRE'] == 'SAN SEBASTIÁN,IGUELDO')&
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 5) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '1024E', 'NOMBRE'] = 'SAN SEBASTIÁN,IGUELDO'

dias = [8,9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1249X') &
                                                    (AEMET_Realidad['LONGITUD'] == 552262) &
                                                    (AEMET_Realidad['LATITUD'] == 432112) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1249X') &
                                                    (AEMET_Realidad['LONGITUD'] == 552262) &
                                                    (AEMET_Realidad['LATITUD'] == 432112) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 8) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '1249X', ['LATITUD', 'LONGITUD']] = [432112, 552262]

dias = [8,9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '3168D') &
                                                    (AEMET_Realidad['LONGITUD'] == 309012) &
                                                    (AEMET_Realidad['LATITUD'] == 403749) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '3168D') &
                                                    (AEMET_Realidad['LONGITUD'] == 309012) &
                                                    (AEMET_Realidad['LATITUD'] == 403749) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 8) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '3168D', ['LATITUD', 'LONGITUD']] = [403749, 309002]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '3260B') &
                                                    (AEMET_Realidad['LONGITUD'] == 402432) &
                                                    (AEMET_Realidad['LATITUD'] == 398800)].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '4642E', ['LATITUD', 'LONGITUD']] = [371642, 654422]

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '5270B', ['LONGITUD', 'LATITUD']] = [348322, 374639]

dias = [2,3,4,5,6]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '5530E') &
                                                    (AEMET_Realidad['LONGITUD'] == 347232) &
                                                    (AEMET_Realidad['LATITUD'] == 371125) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '5530E', ['LATITUD', 'LONGITUD']] = [371125, 347232]

dias = [26,27,28]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '5790Y') &
                                                    (AEMET_Realidad['LONGITUD'] == 600162) &
                                                    (AEMET_Realidad['LATITUD'] == 372158) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '5790Y', ['LATITUD', 'LONGITUD']] = [372151, 600212]

dias = [8,9,10,11,12]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '8368U') &
                                                    (AEMET_Realidad['LONGITUD'] == 107272) &
                                                    (AEMET_Realidad['LATITUD'] == 402103) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7) &
                                                    (AEMET_Realidad['DIA'].isin(dias))].index)

dias1 = [9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '8368U') &
                                                    (AEMET_Realidad['LONGITUD'] == 107272) &
                                                    (AEMET_Realidad['LATITUD'] == 402103) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 8) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '8368U', ['LATITUD', 'LONGITUD']] = [402102, 107272]

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '8500A', ['LATITUD', 'LONGITUD']] = [395726, 4192]
AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '8500A', 'NOMBRE'] = 'CASTELLÓN-ALMAZORA'

dias1 = [28,29,30]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9091R') &
                                                    (AEMET_Realidad['LONGITUD'] == 244002) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 11) &
                                                    (AEMET_Realidad['DIA'].isin(dias1))].index)

dias2 = [1,2]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9091R') &
                                                    (AEMET_Realidad['LONGITUD'] == 244002) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 12) &
                                                    (AEMET_Realidad['DIA'].isin(dias2))].index)

dias3 = [15,16,17,18,19,20]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9091R') &
                                                    (AEMET_Realidad['LONGITUD'] == 243592) &
                                                    (AEMET_Realidad['ANYO'] == 2018) &
                                                    (AEMET_Realidad['MES'] == 5) &
                                                    (AEMET_Realidad['DIA'].isin(dias3))].index)

dias5 = [11,12,13,14,15]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9091R') &
                                                    (AEMET_Realidad['NOMBRE'] == 'VITORIA-GASTEIZ AEROPUERTO') &
                                                    (AEMET_Realidad['LONGITUD'] == 243582) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 5) &
                                                    (AEMET_Realidad['DIA'].isin(dias5))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9091R', ['LATITUD', 'LONGITUD']] = [425219, 243582]
AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9091R', 'NOMBRE'] = 'VITORIA-GASTEIZ AEROPUERTO'

dias5 = [9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9263D') &
                                                    (AEMET_Realidad['NOMBRE'] == 'PAMPLONA, AEROPUERTO') &
                                                    (AEMET_Realidad['ANYO'] == 2021) &
                                                    (AEMET_Realidad['MES'] == 4) &
                                                    (AEMET_Realidad['DIA'].isin(dias5))].index)


AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9263D', 'NOMBRE'] = 'PAMPLONA, AEROPUERTO'

dias5 = [12,13,14,15,16]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9434P') &
                                                    (AEMET_Realidad['NOMBRE'] == 'ZARAGOZA, VALDESPARTERA') &
                                                    (AEMET_Realidad['LONGITUD'] == 56062)&
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 9) &
                                                    (AEMET_Realidad['DIA'].isin(dias5))].index)

dias6 = [9,10,11,12,13]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '9434P') &
                                                    (AEMET_Realidad['NOMBRE'] == 'ZARAGOZA-VALDESPARTERA') &
                                                    (AEMET_Realidad['ANYO'] == 2021) &
                                                    (AEMET_Realidad['MES'] == 4) &
                                                    (AEMET_Realidad['DIA'].isin(dias6))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9434P', ['LATITUD', 'LONGITUD']] = [413716, 56052]
AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == '9434P', 'NOMBRE'] = 'ZARAGOZA, VALDESPARTERA'

dias6 = [2,3,4,5,6]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == 'B228') &
                                                    (AEMET_Realidad['NOMBRE'] == 'PALMA-PUERTO') &
                                                    (AEMET_Realidad['ANYO'] == 2020) &
                                                    (AEMET_Realidad['MES'] == 10) &
                                                    (AEMET_Realidad['DIA'].isin(dias6))].index)

AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == 'B228', ['LATITUD', 'LONGITUD']] = [393319, 237311]
AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == 'B228', 'NOMBRE'] = 'PALMA-PUERTO'

dias6 = [1,2,3,4]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == 'C649I') &
                                                    (AEMET_Realidad['LONGITUD'] == 1523222) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 10) &
                                                    (AEMET_Realidad['DIA'].isin(dias6))].index)

dias7 = [29,30]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == 'C649I') &
                                                    (AEMET_Realidad['LONGITUD'] == 1523222) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 9) &
                                                    (AEMET_Realidad['DIA'].isin(dias7))].index)


AEMET_Realidad.loc[AEMET_Realidad['INDICATIVO'] == 'C649I', ['LATITUD', 'LONGITUD']] = [275504, 1523432]

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['INDICATIVO'] == '1495') &
                                                    (AEMET_Realidad['LONGITUD'] == 837262) &
                                                    (AEMET_Realidad['LATITUD'].isnull())].index)

________________________________________________________________________________

AEMET_Realidad = AEMET_Realidad.drop_duplicates()

dia = [26,27,28,29,30,31]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'REUS/AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 68) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 1) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'REUS/AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 68) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['DIA'] == 1)].index)

dia = [26,27,28]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'TARRAGONA, FAC. GEOGRAFÍA') &
                                                    (AEMET_Realidad['ALTITUD'] == 55) &
                                                    (AEMET_Realidad['ANYO'] == 2015)&
                                                    (AEMET_Realidad['MES'] == 1) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

dia = [26,27,28,29,30,31]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'BARCELONA/AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 4) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 1) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

dia1 = [1,2]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'BARCELONA/AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 4) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['DIA'].isin(dia1))].index)

dia = [26,27,28]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'BARCELONA,CMT') &
                                                    (AEMET_Realidad['ALTITUD'] == 25) &
                                                    (AEMET_Realidad['ANYO'] == 2015) &
                                                    (AEMET_Realidad['MES'] == 1) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

dia = [9,10,11,12,13]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'SAN SEBASTIÁN,IGUELDO') &
                                                    (AEMET_Realidad['ALTITUD'] == 250) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 8) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

dia = [8,9,10,11,12,13]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'SANTANDER,CMT') &
                                                    (AEMET_Realidad['ALTITUD'] == 51) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

dia = [7,8,9,10,11]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'OVIEDO') &
                                                    (AEMET_Realidad['ALTITUD'] == 338) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 2) &
                                                    (AEMET_Realidad['DIA'].isin(dia))].index)

mes = [6,7]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'A CORUÑA') &
                                                    (AEMET_Realidad['ALTITUD'] == 57) &
                                                    (AEMET_Realidad['MES'].isin(mes)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

dia = [8,9,10,11,12,13]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'A CORUÑA') &
                                                    (AEMET_Realidad['ALTITUD'] == 57) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

mes = [10,11]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'VIGO/PEINADOR') &
                                                    (AEMET_Realidad['ALTITUD'] == 255) &
                                                    (AEMET_Realidad['MES'].isin(mes)) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

mes = [6,7]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'PONFERRADA') &
                                                    (AEMET_Realidad['ALTITUD'] == 534) &
                                                    (AEMET_Realidad['MES'].isin(mes)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

mes = [6,7]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'TOLEDO') &
                                                    (AEMET_Realidad['ALTITUD'] == 513) &
                                                    (AEMET_Realidad['MES'].isin(mes)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

dia = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'TOLEDO') &
                                                    (AEMET_Realidad['ALTITUD'] == 513) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [8,9,10,11,12,13]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'PONTEVEDRA') &
                                                    (AEMET_Realidad['ALTITUD'] == 113) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [8,9,10,11,12,13]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'OURENSE') &
                                                    (AEMET_Realidad['ALTITUD'] == 143) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [8,9,10,11,12,13]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'SORIA') &
                                                    (AEMET_Realidad['ALTITUD'] == 1081) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'SEGOVIA') &
                                                    (AEMET_Realidad['ALTITUD'] == 1005) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

mes = [6,7]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CIUDAD REAL') &
                                                    (AEMET_Realidad['ALTITUD'] == 626) &
                                                    (AEMET_Realidad['MES'].isin(mes)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

dia = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CIUDAD REAL') &
                                                    (AEMET_Realidad['ALTITUD'] == 626) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'VALLADOLID') &
                                                    (AEMET_Realidad['ALTITUD'] == 734) &
                                                    (AEMET_Realidad['DIA'] == 8) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'BADAJOZ/TALAVERA LA REAL') &
                                                    (AEMET_Realidad['ALTITUD'] == 185) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [25,26,27,28,29,30]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'BADAJOZ/TALAVERA LA REAL') &
                                                    (AEMET_Realidad['ALTITUD'] == 185) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 11)].index)

dia = [14,15,16,17,18]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'HUELVA, RONDA ESTE') &
                                                    (AEMET_Realidad['ALTITUD'] == 18) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MES'] == 10)].index)

dia = [25,26,27,28,29,30]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'GRANADA/AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 560) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 11)].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CADIZ,OBS.') &
                                                    (AEMET_Realidad['ALTITUD'] == 2) &
                                                    (AEMET_Realidad['DIA'] == 1) &
                                                    (AEMET_Realidad['ANYO'] == 2011) &
                                                    (AEMET_Realidad['MAX'].isnull()) &
                                                    (AEMET_Realidad['MES'] == 3)].index)

dia = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'MURCIA') &
                                                    (AEMET_Realidad['ALTITUD'] == 61) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CUENCA') &
                                                    (AEMET_Realidad['ALTITUD'] == 948) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'ALBACETE,OBS.') &
                                                    (AEMET_Realidad['ALTITUD'] == 676) &
                                                    (AEMET_Realidad['MES'] == 6) &
                                                    (AEMET_Realidad['ANYO'] == 2018) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

ANIO = [2009,2010,2011]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'VITORIA-GASTEIZ AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 513) &
                                                    (AEMET_Realidad['MES'] == 3) &
                                                    (AEMET_Realidad['DIA'] == 1) &
                                                    (AEMET_Realidad['ANYO'].isin(ANIO)) &
                                                    (AEMET_Realidad['MAX'].isnull())].index)

dia = [14,15,16,17,18,19]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'ALBACETE,OBS.') &
                                                    (AEMET_Realidad['ALTITUD'] == 676) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2018) &
                                                    (AEMET_Realidad['MES'] == 6)].index)

dia = [15,16,17,18,19]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'ALBACETE,OBS.') &
                                                    (AEMET_Realidad['ALTITUD'] == 676) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MES'] == 10)].index)

dia = [27,28,29,30]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CASTELLÓN-ALMAZORA') &
                                                    (AEMET_Realidad['ALTITUD'] == 43) &
                                                    (AEMET_Realidad['DIA'].isin(dia)) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MES'] == 4)].index)

dia1 = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CASTELLÓN-ALMAZORA') &
                                                    (AEMET_Realidad['ALTITUD'] == 43) &
                                                    (AEMET_Realidad['DIA'].isin(dia1)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'CASTELLÓN-ALMAZORA') &
                                                    (AEMET_Realidad['ALTITUD'] == 43) &
                                                    (AEMET_Realidad['DIA'] == 1) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MES'] == 5)].index)

dia1 = [8,9,10,11,12]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'LLEIDA') &
                                                    (AEMET_Realidad['ALTITUD'] == 185) &
                                                    (AEMET_Realidad['DIA'].isin(dia1)) &
                                                    (AEMET_Realidad['ANYO'] == 2023) &
                                                    (AEMET_Realidad['MES'] == 7)].index)

dia1 = [1,2,3,4,5]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'HUESCA, AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 541) &
                                                    (AEMET_Realidad['DIA'].isin(dia1)) &
                                                    (AEMET_Realidad['ANYO'] == 2016) &
                                                    (AEMET_Realidad['MES'] == 11)].index)

dia1 = [9,10,11,12,13,14]
AEMET_Realidad = AEMET_Realidad.drop(AEMET_Realidad[(AEMET_Realidad['NOMBRE'] == 'GRAN CANARIA/AEROPUERTO') &
                                                    (AEMET_Realidad['ALTITUD'] == 24) &
                                                    (AEMET_Realidad['DIA'].isin(dia1)) &
                                                    (AEMET_Realidad['ANYO'] == 2022) &
                                                    (AEMET_Realidad['MES'] == 11)].index)

# Calcular el porcentaje de valores vacíos por columna
porcentaje_perdidos = AEMET_Realidad.isnull().mean() * 100
porcentaje_perdidos_act = AEMET_Realidad.isnull().mean() * 100

# Mostrar el porcentaje de valores vacíos por columna
print(porcentaje_perdidos)
print()
print(porcentaje_perdidos_act)

AEMET_Realidad.to_csv(path + 'AEMET_Realidad_limpio.csv', index=False)

Prediccion_AEMET_corta = AEMET_Previsiones[AEMET_Previsiones['FECHA_FICH'] == AEMET_Previsiones['FECHA']]
Prediccion_AEMET_corta['FECHA'] = pd.to_datetime(Prediccion_AEMET_corta['FECHA'], format='%d/%m/%Y')

print("DataFrame Filtrado:")
print(Prediccion_AEMET_corta.head(15))

# Hacer la combinación de los DataFrames
merged_df = AEMET_Realidad.merge(Prediccion_AEMET_corta, left_on=['FECHA', 'INDICATIVO'], right_on=['FECHA', 'ID_ESTACION'], how='left')

# Rellenar los valores perdidos en MAX y MIN
merged_df['MAX'] = merged_df['MAX'].combine_first(merged_df['T_MAX'])
merged_df['MIN'] = merged_df['MIN'].combine_first(merged_df['T_MIN'])

# Seleccionar solo las columnas originales de AEMET_Realidad
AEMET_Realidad_actualizado = merged_df[AEMET_Realidad.columns]

print("DataFrame AEMET_Realidad Actualizado:")
print(AEMET_Realidad_actualizado)

#creamos una funcion para colocar el signo correcto a las Longitudes y Latitudes
def ajustar_longitud(df):
    df=df.copy()
    # Diccionario con provincias y si están al oeste del meridiano
    provincias_oeste = {
        'TARRAGONA': False, 'BARCELONA': False, 'GIRONA': False, 'GIPUZKOA': True,
        'BIZKAIA': True, 'CANTABRIA': True, 'ASTURIAS': True, 'A CORUÑA': True,
        'PONTEVEDRA': True, 'LUGO': True, 'LEON': True, 'OURENSE': True,
        'SORIA': True, 'BURGOS': True, 'PALENCIA': True, 'VALLADOLID': True,
        'AVILA': True, 'SEGOVIA': True, 'ZAMORA': True, 'SALAMANCA': True,
        'MADRID': True, 'GUADALAJARA': True, 'TOLEDO': True, 'CACERES': True,
        'CIUDAD REAL': True, 'BADAJOZ': True, 'HUELVA': True, 'JAEN': True,
        'CORDOBA': True, 'GRANADA': True, 'SEVILLA': True, 'CADIZ': True,
        'MALAGA': True, 'ALMERIA': True, 'MURCIA': False, 'ALICANTE': False,
        'CUENCA': True, 'ALBACETE': True, 'TERUEL': False, 'VALENCIA': False,
        'CASTELLON': False, 'ARABA/ALAVA': True, 'LA RIOJA': True, 'NAVARRA': True,
        'ZARAGOZA': False, 'LLEIDA': False, 'HUESCA': False, 'BALEARES': False,
        'LAS PALMAS': True
    }
    
    # Aplicar el ajuste de longitud
    df['LONGITUD'] = df.apply(lambda row: -row['LONGITUD'] if provincias_oeste.get(row['NOM_PROV'], False) else row['LONGITUD'], axis=1)
    return df

#creamos una funcion para corregir las posiciones de las latitudes y longitudes
def ajustar_formato_longitud_latitud(df):
    df=df.copy()
    def convertir_longitud(valor):
        if pd.isnull(valor) or not isinstance(valor, (int, float, str)):
            return None
        try:
            valor_str = str(valor).strip().replace('.', '')
            if len(valor_str) < 6:
                return None
            parte_entera = valor_str[:-6]
            parte_decimal = valor_str[-6:]
            return float(f"{parte_entera}.{parte_decimal}")
        except ValueError:
            return None

    def convertir_latitud(valor):
        if pd.isnull(valor) or not isinstance(valor, (int, float, str)):
            return None
        try:
            valor_str = str(valor).strip().replace('.', '')
            if len(valor_str) < 6:
                return None
            parte_entera = valor_str[:-5]
            parte_decimal = valor_str[-5:]
            return float(f"{parte_entera}.{parte_decimal}")
        except ValueError:
            return None

    df['LONGITUD'] = df['LONGITUD'].apply(convertir_longitud)
    df['LATITUD'] = df['LATITUD'].apply(convertir_latitud)
    
    return df 

#creamos un formato mas amigable para las busquedas por coordinadas
def convertir_a_formato_busqueda(df):
    df=df.copy()
    # Crear la columna 'GOOGLE_MAPS' con el formato 'latitud,longitud'
    df['GOOGLE_MAPS'] = df.apply(lambda row: f"{row['LATITUD']},{row['LONGITUD']}", axis=1)
    return df

AEMET_Realidad_actualizado = ajustar_longitud(AEMET_Realidad_actualizado)
AEMET_Realidad_actualizado = ajustar_formato_longitud_latitud(AEMET_Realidad_actualizado)
AEMET_Realidad_actualizado = convertir_a_formato_busqueda(AEMET_Realidad_actualizado)
AEMET_Realidad_actualizado.head()

AEMET_Realidad_actualizado.to_csv(path + '\\Cleaning\\AEMET_Realidad_clean.csv', index=False)

# Calcular el porcentaje de valores vacíos por columna
porcentaje_perdidos = AEMET_Realidad_actualizado.isnull().mean() * 100
porcentaje_perdidos_act = AEMET_Realidad_actualizado.isnull().mean() * 100

# Mostrar el porcentaje de valores vacíos por columna
print(porcentaje_perdidos)
print()
print(porcentaje_perdidos_act)

def calcular_temperaturas_promedio_por_provincia_y_dia(df):
    """
    Calcula las temperaturas máximas y mínimas por día, promediando por provincia.

    Parámetros:
    df (pandas.DataFrame): El DataFrame que contiene las columnas 'FECHA', 'NOM_PROV', 'MAX', y 'MIN'.

    Retorna:
    pandas.DataFrame: Un DataFrame con las temperaturas promedio por provincia y día.
    """
    # Asegurarse de que la columna FECHA esté en formato de fecha
    df['FECHA'] = pd.to_datetime(df['FECHA'], format='%Y-%m-%d')

    # Agrupar por NOM_PROV y FECHA para calcular el promedio de MAX y MIN
    temperatura_promedio = df.groupby(['NOM_PROV', 'FECHA']).agg({
        'MAX': 'mean',
        'MIN': 'mean'
    }).reset_index()

    # Redondear las temperaturas a dos decimales
    temperatura_promedio['MAX'] = temperatura_promedio['MAX'].round(2)
    temperatura_promedio['MIN'] = temperatura_promedio['MIN'].round(2)

    return temperatura_promedio

# Uso de la función con el DataFrame AEMET_Realidad_actualizado
provincias_temperaturas_promedio = calcular_temperaturas_promedio_por_provincia_y_dia(AEMET_Realidad_actualizado)
# Ordenar el DataFrame por la columna FECHA
provincias_temperaturas_promedio = provincias_temperaturas_promedio.sort_values(by='FECHA')
# Verificar el resultado
provincias_temperaturas_promedio.head(3)

#Guardamos nuestro dataset por si lo necesitamos mas adelante
provincias_temperaturas_promedio.to_csv(path + '\\Cleaning\\provincias_temperaturas_promedio.csv', index=False)
