==> Notebooks/2_Clean_Festivos.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
from astral.sun import sun
from astral import LocationInfo

# Especifica la ruta al archivo CSV
path = r'C:\\Users\\joset\\OneDrive\\Documentos\\Master Big Data UCM\\18-TFN\NaturgyProject\\data'
# Lee el archivo CSV en un DataFrame de pandas
Festivos_Principales_original = pd.read_csv(path + '\\Originales\\Festivos_Principales.csv', encoding = 'latin1')


Festivos_Principales = Festivos_Principales_original.copy()

Festivos_Principales.head(3)

Festivos_Principales.columns

def convertir_fecha(fecha):
    return pd.to_datetime(fecha.replace('y', '').replace('m', '-').replace('d', '-'), format='%Y-%m-%d')


# Aplicar la función a la columna 'Diario'
Festivos_Principales['Diario'] = Festivos_Principales['Diario'].apply(convertir_fecha)
Festivos_Principales.head(3)

# Selecciono las columnas que quiero eliminar para aligerar mi dataframe, no me interesan los festivos comunitarios
columnas_eliminar = ['Fest_ComMadrid', 'Fest_SanJuan',
       'Fest_Santiago', 'Fest_DiaCatalunya', 'Fest_Merce', 'Fest_SanIsidro']
Festivos_Principales.drop(columns=columnas_eliminar, inplace=True, axis = 2)

#vemos como queda los festivos quitando las festividades de la comunidad que tienen menos importancia
Festivos_Principales.columns

#diccionario de Feriados:
columnas_festivos = [
    'Fest_AnhoNuevo', 'Fest_Reyes', 'Fest_SatSemSantaLMX',
       'Fest_JuevesSanto', 'Fest_ViernesSanto', 'Fest_Trabajo',
       'Fest_VirgenAgosto', 'Fest_Hispanidad', 'Fest_TodosLosSantos',
       'Fest_Constitucion', 'Fest_Inmaculada', 'Fest_Nochebuena',
       'Fest_Navidad', 'Fest_Findeanho'
]


Festivos_Espana = pd.DataFrame({
    'FECHA': Festivos_Principales['Diario'],
    'FESTIVO': 0
})

for columna in columnas_festivos:
    Festivos_Espana['FESTIVO'] = Festivos_Espana['FESTIVO'] | Festivos_Principales[columna]
Festivos_Espana['DIAS'] = Festivos_Espana['FECHA'] .dt.day_name()
Festivos_Espana['SEMANA'] = Festivos_Espana['FECHA'].dt.isocalendar().week

Festivos_Espana.head(3)

# Función para determinar la estación del año en Europa
def estacion_del_ano(fecha):
    mes = fecha.month
    dia = fecha.day
    
    if (mes == 12 and dia >= 21) or mes in [1, 2] or (mes == 3 and dia < 21):
        return 1  # Invierno
    elif (mes == 3 and dia >= 21) or mes in [4, 5] or (mes == 6 and dia < 21):
        return 2  # Primavera
    elif (mes == 6 and dia >= 21) or mes in [7, 8] or (mes == 9 and dia < 21):
        return 3  # Verano
    else:
        return 4  # Otoño


# Crear un objeto de ubicación para Madrid, España
madrid = LocationInfo("Madrid", "Spain", "Europe/Madrid", 40.4168, -3.7038)
# Función para calcular las horas de luz usando la biblioteca astral
def horas_de_luz(fecha):
    s = sun(madrid.observer, date=fecha)
    amanecer = s['sunrise']
    atardecer = s['sunset']
    
    # Calcular la duración de la luz del día en horas
    duracion_luz = (atardecer - amanecer).seconds / 3600.0
    
    return round(duracion_luz, 2)

# Aplicar la función a la columna 'FECHA' para agregar la estación del año
Festivos_Espana['Estacion'] = Festivos_Espana['FECHA'].apply(estacion_del_ano)
# Aplicar la función a la columna 'FECHA' para agregar las horas de luz estimadas
Festivos_Espana['Horas_Luz'] = Festivos_Espana['FECHA'].apply(horas_de_luz)

Festivos_Espana.head(3)

#Guardamos nuestro dataset por si lo necesitamos mas adelante
Festivos_Espana.to_csv(path + '\\Cleaning\\Festivos_Principales_clean.csv', index=False)
